requirements.txt
mediapipe
streamlit
streamlit-webrtc
av
pillow